requirements.txt
requests==2.31.0
kafka-python==2.0.2
redis==5.0.1
psycopg2-binary==2.9.9
